requirements.txt
discord.py
github
ShitDB
